OpenMP Parallel Implementation Added OpenMP support for the main loop of the complex source generation Might remove the simple source algorithm as it is no longer used Changed the random number generator for the entire simulation to be thread safe Changed the count incrementation of each image plate to be done by reference instead of using the external symbols Will also profile and optimize some parts of the code that seem very "dirty"

+++ simulation_interface.h
@@ -32,12 +32,12 @@ public:
     
     struct Stats
     {
-        double c_sour;
-        double c_cr1;
-        double c_cr2_para;
-        double c_cr2_anti;
-        double c_detc_para;
-        double c_detc_anti;
+        int c_sour;
+        int c_cr1;
+        int c_cr2_para;
+        int c_cr2_anti;
+        int c_detc_para;
+        int c_detc_anti;
         double delrot;
         std::vector<std::vector<double>> events_para;
         std::vector<std::vector<double>> events_anti;
